feature: create module-level config file.
Adds the config module which facilitates managing configuration for pylabrobot. Default config for development is in pylabrobot.ini which mimics the config that was already found in the module.

Configs are searched for in the current directory and all parent directory. If none are found, a new one is created in the current working directory.

Currently, it supports .ini files, but wrote to easily support json, yaml, etc.

I set the default log level to INFO for end users.