feat: Add support for using ResourceStack with PlateCarrier
This commit modifies the `assign_child_resource` method in the `PlateCarrierSite` class to allow the storage of `ResourceStack` objects in addition to `Plate` and `PlateAdapter` resources. The code now checks that the top resource in the `ResourceStack` is a `Plate`, `PlateAdapter`, or `Lid` before assigning it to the `PlateCarrierSite`. This change enhances the flexibility of the `PlateCarrierSite` and expands its capabilities.